s/config profile/profile
Closes #615
