+++ specs/configuration.md
@@ -120,7 +120,7 @@ the whole entry is overriden by the new definition.
 
 Device entries are added through:
     lxc config device add <container> <name> <type> [key=value]...
-    lxc config profile device add <profile> <name> <type> [key=value]...
+    lxc profile device add <profile> <name> <type> [key=value]...
 
 ## Profiles
 Profiles can store any configuration that a container can (key/value or devices)
